Explicitly convert size_t to int
Bug: None

+++ call/call.cc
@@ -1323,7 +1323,7 @@ PacketReceiver::DeliveryStatus Call::DeliverRtcp(MediaType media_type,
 PacketReceiver::DeliveryStatus Call::DeliverRtp(MediaType media_type,
                                                 rtc::CopyOnWriteBuffer packet,
                                                 const PacketTime& packet_time) {
-  int length = packet.size();
+  int length = static_cast<int>(packet.size());
   TRACE_EVENT0("webrtc", "Call::DeliverRtp");
 
   RtpPacketReceived parsed_packet;
